ShoppingCart deleted when Order consumed

+++ src/main/java/com/rsr/shopping_cart_microservice/core/domain/service/impl/ShoppingCartService.java
@@ -183,8 +183,10 @@ public class ShoppingCartService implements IShoppingCartService {
 
     @Override
     public void deleteCart(UUID userId) throws NoCartException {
-        ShoppingCart cart = shoppingCartRepository.findById(userId)
-                .orElseThrow(() -> new NoCartException(userId));
+        ShoppingCart cart = shoppingCartRepository.findByUserId(userId);
+        if (cart == null) {
+            throw new NoCartException(userId);
+        }
         shoppingCartRepository.delete(cart);
     }
